requirements.txt
requests
transformers
torch
Pillow
numpy
joblib
shap
